Add files via upload
This repo is created just for learning purposes. Image is directly passed from the algorithm, then tried with few features extractors which gives inconsistent shapes of features and finally HOG feature extractor is used to obtained the accurate shape of features to be passed to model. also time complexity is examined.